[Accelerator] Remove static NNPA libs by resolving circular dependencies (#1269)
* initial code


* Do not use static libs


* clean up


* Remove more static libs


* Minor changes


* Address comments


* just to trigger CI tests


* Address comments about cmake


* Move some options from NNPAAccelerator.cpp into NNPACompilerUtils.cpp
