+++ src/Accelerators/NNPA/Compiler/CMakeLists.txt
@@ -1,8 +1,9 @@
 get_property(OMLibs GLOBAL PROPERTY ONNX_MLIR_LIBS)
 
-add_onnx_mlir_library(NNPACompilerUtils STATIC
-  EXCLUDE_FROM_OM_LIBS
+add_onnx_mlir_library(NNPACompilerUtils
   NNPACompilerUtils.cpp
+
+  EXCLUDE_FROM_OM_LIBS
 
   INCLUDE_DIRS PRIVATE
   ${NNPA_SRC_ROOT}
@@ -11,8 +12,8 @@ add_onnx_mlir_library(NNPACompilerUtils STATIC
   ${NNPA_ONNX_MLIR_BIN_ROOT}
 
   LINK_LIBS PUBLIC
-  CompilerUtils
-  #${OMLibs}
+  ${OMLibs}
+  CompilerPasses
 
   ACCEL_INCLUDE_DIRS PRIVATE
   ${NNPA_ONNX_MLIR_SRC_ROOT}
