+++ src/Accelerators/NNPA/Compiler/NNPACompilerUtils.hpp
@@ -13,9 +13,10 @@
 
 #pragma once
 
+#include "mlir/IR/BuiltinOps.h"
 #include "mlir/Pass/PassManager.h"
-#include "src/Compiler/CompilerUtils.hpp"
-#include "src/Support/OMOptions.hpp"
+#include "mlir/Pass/PassRegistry.h"
+#include "onnx-mlir/Compiler/OMCompilerTypes.h"
 
 namespace onnx_mlir {
 
@@ -34,15 +35,6 @@ void addZHighToZLowPasses(mlir::PassManager &pm);
 void addAllToLLVMPasses(mlir::PassManager &pm);
 
 void addPassesNNPA(mlir::OwningOpRef<mlir::ModuleOp> &module,
-    mlir::PassManager &pm, onnx_mlir::EmissionTargetType &emissionTarget,
-    NNPAEmissionTargetType nnpaEmissionTarget,
-    mlir::ArrayRef<std::string> execNodesOnCpu);
-
-int compileModuleNNPA(mlir::OwningOpRef<mlir::ModuleOp> &module,
-    mlir::MLIRContext &context, std::string outputBaseName,
-    onnx_mlir::EmissionTargetType emissionTarget,
-    NNPAEmissionTargetType dlcEmissionTarget =
-        NNPAEmissionTargetType::EmitZNONE,
-    mlir::ArrayRef<std::string> execNodesOnCpu = mlir::ArrayRef<std::string>());
+    mlir::PassManager &pm, onnx_mlir::EmissionTargetType &emissionTarget);
 
 } // namespace onnx_mlir
