+++ src/Compiler/CMakeLists.txt
@@ -24,10 +24,48 @@ add_custom_target(ExternalUtil DEPENDS ${FILE_GENERATE_DIR}/ExternalUtil.hpp)
 
 get_property(OMLibs GLOBAL PROPERTY ONNX_MLIR_LIBS)
 
-add_onnx_mlir_library(CompilerUtils STATIC
-  CompilerUtils.cpp
+add_onnx_mlir_library(CompilerOptions
+  CompilerOptions.cpp
 
   EXCLUDE_FROM_OM_LIBS
+
+  DEPENDS
+  ExternalUtil
+
+  INCLUDE_DIRS PRIVATE
+  ${FILE_GENERATE_DIR}
+  
+  INCLUDE_DIRS PUBLIC
+  ${ONNX_MLIR_SRC_ROOT}/include
+
+  LINK_LIBS PUBLIC
+  MLIRIR
+  )
+
+add_onnx_mlir_library(CompilerPasses
+  CompilerPasses.cpp
+
+  EXCLUDE_FROM_OM_LIBS
+
+  DEPENDS
+  MLIRIR
+
+  INCLUDE_DIRS PRIVATE
+  ${FILE_GENERATE_DIR}
+  
+  INCLUDE_DIRS PUBLIC
+  ${ONNX_MLIR_SRC_ROOT}/include
+
+  LINK_LIBS PUBLIC
+  ${OMLibs}
+  CompilerOptions
+  MLIRAffineTransforms
+  MLIRLinalgTransforms
+  MLIRLLVMToLLVMIRTranslation
+  )
+
+add_onnx_mlir_library(CompilerUtils
+  CompilerUtils.cpp
 
   DEPENDS
   ExternalUtil
@@ -42,9 +80,8 @@ add_onnx_mlir_library(CompilerUtils STATIC
 
   LINK_LIBS PUBLIC
   ${OMLibs}
-  MLIRAffineTransforms
-  MLIRLinalgTransforms
-  MLIRLLVMToLLVMIRTranslation
+  CompilerPasses
+  InitAccelerators
 
   # Link LLVM libraries necessary to query which target architectures are configured.
   LINK_COMPONENTS PRIVATE
