+++ src/Tools/onnx-mlir-opt/CMakeLists.txt
@@ -7,6 +7,8 @@ add_onnx_mlir_executable(onnx-mlir-opt
 
   LINK_LIBS PRIVATE
   ${OMLibs}
+  CompilerOptions
+  InitAccelerators
   MLIRAffineTransforms
   MLIRLinalgTransforms
   MLIRMemRefTransforms
