+++ src/Tools/onnx-mlir-opt/onnx-mlir-opt.cpp
@@ -30,12 +30,12 @@
 #include <mlir/Support/MlirOptMain.h>
 
 #include "src/Accelerators/Accelerator.hpp"
+#include "src/Compiler/CompilerOptions.hpp"
 #include "src/Dialect/Krnl/KrnlOps.hpp"
 #include "src/Dialect/ONNX/ONNXOps.hpp"
 #include "src/InitMLIRPasses.hpp"
 #include "src/InitOMPasses.hpp"
 #include "src/Pass/Passes.hpp"
-#include "src/Support/OMOptions.hpp"
 
 using namespace mlir;
 using namespace onnx_mlir;
